This PR:
- Incorporates feedback from the previous PR
- Calls out some obsoleted functionality in Filters. This will be removed in a future PR.
- Adds FilteredValueAssignmentContextTests
- Tidies some behaviour to machine/surveyed surface arrays being null vs Guid[0]
- Spreads use of VersionSerializationHelper through all From/ToBinary contexts.
- Added additional version serializer support and tests for IBinaryReaderWriter style contexts.
- Refactored TRexGridFactory DI to allow consistent use from unit tests.
- Fixes issue where BaseRerquest derived TRex requests were not calling the BaseRequest ToBinary/FromBinary ancestor virtual method.

Related work items: #80650

+++ src/service/TRex/src/netstandard/VSS.TRex.SubGrids/GridRotationUtilities.cs
@@ -252,7 +252,7 @@ namespace VSS.TRex.SubGrids
           {
             sieveBitmask.SetBit(eastCol, northRow);
             assignmentContext.ProbePositions[eastCol, northRow]
-              .SetOffsets(currentEast - subGridMinX, currentNorth - subGridMinY);
+              .SetOffsets((float)(currentEast - subGridMinX), (float)(currentNorth - subGridMinY));
           }
 
           currentEast += stepEastX;
